+++ timeslip/status/status.go
@@ -1,17 +1,9 @@
 // Package status contains the basic states of a timeslip.
 package status
 
-// Started status string.
-func Started() string {
-	return "started"
-}
-
-// Paused status string.
-func Paused() string {
-	return "paused"
-}
-
-// Completed status string.
-func Completed() string {
-	return "completed"
-}
+const (
+	Started   = "started"
+	Paused    = "paused"
+	Resumed   = "resumed"
+	Completed = "completed"
+)